@@ -1,29 +1,24 @@
 {
-  "name": "weather",
-  "version": "0.1.0",
+  "name": "weather-app",
+  "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "axios": "^0.19.2",
-    "react": "^16.13.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
     "react-animated-weather": "^4.0.1",
-    "react-dom": "^16.13.1",
-    "react-geolocated": "^3.0.1",
-    "react-live-clock": "^4.0.5",
-    "react-scripts": "3.4.1",
-    "react-skycons": "^0.7.0",
-    "skycons": "^1.0.0"
+    "react-live-clock": "^6.1.21"
   },
   "scripts": {
-    "start": "set NODE_OPTIONS=--openssl-legacy-provider && react-scripts start",
-    "build": "set NODE_OPTIONS=--openssl-legacy-provider && react-scripts build",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app"
+    ]
   },
   "browserslist": {
     "production": [
