@@ -1,5 +1,6 @@
-// Visit https://api.openweathermap.org & then signup to get our API keys for free
-module.exports = {
+const apiKeys = {
   key: "7ef8efe69c55fa78d32afcbf5d64cae9",
-  base: "https://api.openweathermap.org/data/2.5/",
+  base: "https://api.openweathermap.org/data/2.5/"
 };
+
+export default apiKeys;
