@@ -1,73 +1,138 @@
-import React from "react";
+import React, { useState, useEffect, useCallback } from "react";
 import apiKeys from "./apiKeys";
 import Clock from "react-live-clock";
-import Forcast from "./forcast";
+import Forecast from "./Forecast";
 import loader from "./images/WeatherIcons.gif";
 import ReactAnimatedWeather from "react-animated-weather";
+
 const dateBuilder = (d) => {
-  let months = [
-    "January",
-    "February",
-    "March",
-    "April",
-    "May",
-    "June",
-    "July",
-    "August",
-    "September",
-    "October",
-    "November",
-    "December",
+  const months = [
+    "January", "February", "March", "April", "May", "June",
+    "July", "August", "September", "October", "November", "December"
   ];
-  let days = [
-    "Sunday",
-    "Monday",
-    "Tuesday",
-    "Wednesday",
-    "Thursday",
-    "Friday",
-    "Saturday",
+  const days = [
+    "Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"
   ];
 
-  let day = days[d.getDay()];
-  let date = d.getDate();
-  let month = months[d.getMonth()];
-  let year = d.getFullYear();
+  const day = days[d.getDay()];
+  const date = d.getDate();
+  const month = months[d.getMonth()];
+  const year = d.getFullYear();
 
   return `${day}, ${date} ${month} ${year}`;
 };
+
 const defaults = {
   color: "white",
   size: 112,
   animate: true,
 };
-class Weather extends React.Component {
-  state = {
-    lat: undefined,
-    lon: undefined,
-    errorMessage: undefined,
-    temperatureC: undefined,
-    temperatureF: undefined,
-    city: undefined,
-    country: undefined,
-    humidity: undefined,
-    description: undefined,
+
+const getWeatherIcon = (weatherMain) => {
+  const iconMap = {
+    "Haze": "CLEAR_DAY",
+    "Clouds": "CLOUDY",
+    "Rain": "RAIN",
+    "Snow": "SNOW",
+    "Dust": "WIND",
+    "Drizzle": "SLEET",
+    "Fog": "FOG",
+    "Smoke": "FOG",
+    "Tornado": "WIND"
+  };
+  return iconMap[weatherMain] || "CLEAR_DAY";
+};
+
+function CurrentLocation() {
+  const [weather, setWeather] = useState({
+    lat: null,
+    lon: null,
+    temperatureC: null,
+    city: null,
+    country: null,
+    humidity: null,
+    main: null,
     icon: "CLEAR_DAY",
-    sunrise: undefined,
-    sunset: undefined,
-    errorMsg: undefined,
+    windSpeed: null,
+    visibility: null
+  });
+
+  const getPosition = () => {
+    return new Promise((resolve, reject) => {
+      navigator.geolocation.getCurrentPosition(resolve, reject);
+    });
   };
 
-  componentDidMount() {
+  const getWeather = useCallback(async (lat, lon) => {
+    try {
+      const response = await fetch(
+        `${apiKeys.base}weather?lat=${lat}&lon=${lon}&units=metric&APPID=${apiKeys.key}`
+      );
+      const data = await response.json();
+      
+      setWeather({
+        lat,
+        lon,
+        city: data.name,
+        temperatureC: Math.round(data.main.temp),
+        humidity: data.main.humidity,
+        main: data.weather[0].main,
+        country: data.sys.country,
+        visibility: data.visibility,
+        windSpeed: data.wind?.speed || null,
+        icon: getWeatherIcon(data.weather[0].main)
+      });
+    } catch (error) {
+      console.error("Error fetching weather:", error);
+    }
+  }, []);
+
+  const searchByCity = async (cityName) => {
+    try {
+      const response = await fetch(
+        `${apiKeys.base}weather?q=${cityName}&units=metric&APPID=${apiKeys.key}`
+      );
+      const data = await response.json();
+      
+      setWeather({
+        lat: data.coord.lat,
+        lon: data.coord.lon,
+        city: data.name,
+        temperatureC: Math.round(data.main.temp),
+        humidity: data.main.humidity,
+        main: data.weather[0].main,
+        country: data.sys.country,
+        visibility: data.visibility,
+        windSpeed: data.wind?.speed || null,
+        icon: getWeatherIcon(data.weather[0].main)
+      });
+    } catch (error) {
+      console.error("Error fetching city weather:", error);
+    }
+  };
+
+  const getCurrentLocation = () => {
     if (navigator.geolocation) {
-      this.getPosition()
-        //If user allow location service then will fetch data & send it to get-weather function.
+      navigator.geolocation.getCurrentPosition(
+        (position) => {
+          getWeather(position.coords.latitude, position.coords.longitude);
+        },
+        (error) => {
+          console.error("Error getting location:", error);
+          getWeather(28.67, 77.22);
+        }
+      );
+    }
+  };
+
+  useEffect(() => {
+    if (navigator.geolocation) {
+      getPosition()
         .then((position) => {
-          this.getWeather(position.coords.latitude, position.coords.longitude);
+          getWeather(position.coords.latitude, position.coords.longitude);
         })
-        .catch((err) => {
-          //If user denied location service then standard location weather will le shown on basis of latitude & latitude.
-          this.getWeather(28.67, 77.22);
+        .catch(() => {
+          getWeather(28.67, 77.22);
           alert(
             "You have disabled location service. Allow 'This APP' to access your location. Your current location will be used for calculating Real time weather."
           );
@@ -76,222 +141,75 @@ class Weather extends React.Component {
       alert("Geolocation not available");
     }
 
-    this.timerID = setInterval(
-      () => this.getWeather(this.state.lat, this.state.lon),
-      600000
-    );
-  }
+    const timerID = setInterval(() => {
+      if (weather.lat && weather.lon) {
+        getWeather(weather.lat, weather.lon);
+      }
+    }, 600000);
 
-  componentWillUnmount() {
-    clearInterval(this.timerID);
-  }
+    return () => clearInterval(timerID);
+  }, [getWeather, weather.lat, weather.lon]);
 
-  // tick = () => {
-  //   this.getPosition()
-  //   .then((position) => {
-  //     this.getWeather(position.coords.latitude, position.coords.longitude)
-  //   })
-  //   .catch((err) => {
-  //     this.setState({ errorMessage: err.message });
-  //   });
-  // }
-
-  getPosition = (options) => {
-    return new Promise(function (resolve, reject) {
-      navigator.geolocation.getCurrentPosition(resolve, reject, options);
-    });
-  };
-  getWeather = async (lat, lon) => {
-    const api_call = await fetch(
-      `${apiKeys.base}weather?lat=${lat}&lon=${lon}&units=metric&APPID=${apiKeys.key}`
-    );
-    const data = await api_call.json();
-    this.setState({
-      lat: lat,
-      lon: lon,
-      city: data.name,
-      temperatureC: Math.round(data.main.temp),
-      temperatureF: Math.round(data.main.temp * 1.8 + 32),
-      humidity: data.main.humidity,
-      main: data.weather[0].main,
-      country: data.sys.country,
-      visibility: data.visibility,
-      windSpeed: data.wind ? data.wind.speed : null,
-      // sunrise: this.getTimeFromUnixTimeStamp(data.sys.sunrise),
-
-      // sunset: this.getTimeFromUnixTimeStamp(data.sys.sunset),
-    });
-    switch (this.state.main) {
-      case "Haze":
-        this.setState({ icon: "CLEAR_DAY" });
-        break;
-      case "Clouds":
-        this.setState({ icon: "CLOUDY" });
-        break;
-      case "Rain":
-        this.setState({ icon: "RAIN" });
-        break;
-      case "Snow":
-        this.setState({ icon: "SNOW" });
-        break;
-      case "Dust":
-        this.setState({ icon: "WIND" });
-        break;
-      case "Drizzle":
-        this.setState({ icon: "SLEET" });
-        break;
-      case "Fog":
-        this.setState({ icon: "FOG" });
-        break;
-      case "Smoke":
-        this.setState({ icon: "FOG" });
-        break;
-      case "Tornado":
-        this.setState({ icon: "WIND" });
-        break;
-      default:
-        this.setState({ icon: "CLEAR_DAY" });
-    }
-  };
-
-  searchByCity = async (cityName) => {
-    try {
-      const api_call = await fetch(
-        `${apiKeys.base}weather?q=${cityName}&units=metric&APPID=${apiKeys.key}`
-      );
-      const data = await api_call.json();
-      this.setState({
-        lat: data.coord.lat,
-        lon: data.coord.lon,
-        city: data.name,
-        temperatureC: Math.round(data.main.temp),
-        temperatureF: Math.round(data.main.temp * 1.8 + 32),
-        humidity: data.main.humidity,
-        main: data.weather[0].main,
-        country: data.sys.country,
-        visibility: data.visibility,
-        windSpeed: data.wind ? data.wind.speed : null,
-      });
-      this.updateWeatherIcon(data.weather[0].main);
-    } catch (error) {
-      console.error("Error fetching city weather:", error);
-    }
-  };
-
-  updateWeatherIcon = (weatherMain) => {
-    switch (weatherMain) {
-      case "Haze":
-        this.setState({ icon: "CLEAR_DAY" });
-        break;
-      case "Clouds":
-        this.setState({ icon: "CLOUDY" });
-        break;
-      case "Rain":
-        this.setState({ icon: "RAIN" });
-        break;
-      case "Snow":
-        this.setState({ icon: "SNOW" });
-        break;
-      case "Dust":
-        this.setState({ icon: "WIND" });
-        break;
-      case "Drizzle":
-        this.setState({ icon: "SLEET" });
-        break;
-      case "Fog":
-        this.setState({ icon: "FOG" });
-        break;
-      case "Smoke":
-        this.setState({ icon: "FOG" });
-        break;
-      case "Tornado":
-        this.setState({ icon: "WIND" });
-        break;
-      default:
-        this.setState({ icon: "CLEAR_DAY" });
-    }
-  };
-
-  getCurrentLocation = () => {
-    if (navigator.geolocation) {
-      navigator.geolocation.getCurrentPosition(
-        (position) => {
-          this.getWeather(position.coords.latitude, position.coords.longitude);
-        },
-        (error) => {
-          console.error("Error getting location:", error);
-          this.getWeather(28.67, 77.22);
-        }
-      );
-    }
-  };
-
-  render() {
-    if (this.state.temperatureC) {
-      return (
-        <React.Fragment>
-          <button onClick={this.getCurrentLocation} className="location-btn">
-            📍
-          </button>
-          <div className="city">
-            <div className="title">
-              <h2>{this.state.city}</h2>
-              <h3>{this.state.country}</h3>
-            </div>
-            <div className="mb-icon">
-              {" "}
-              <ReactAnimatedWeather
-                icon={this.state.icon}
-                color={defaults.color}
-                size={defaults.size}
-                animate={defaults.animate}
-              />
-              <p>{this.state.main}</p>
-            </div>
-            <div className="date-time">
-              <div className="dmy">
-                <div id="txt"></div>
-                <div className="current-time">
-                  <Clock format="HH:mm:ss" interval={1000} ticking={true} />
-                </div>
-                <div className="current-date">{dateBuilder(new Date())}</div>
+  if (weather.temperatureC) {
+    return (
+      <>
+        <button onClick={getCurrentLocation} className="location-btn">
+          📍
+        </button>
+        <div className="city">
+          <div className="title">
+            <h2>{weather.city}</h2>
+            <h3>{weather.country}</h3>
+          </div>
+          <div className="mb-icon">
+            <ReactAnimatedWeather
+              icon={weather.icon}
+              color={defaults.color}
+              size={defaults.size}
+              animate={defaults.animate}
+            />
+            <p>{weather.main}</p>
+          </div>
+          <div className="date-time">
+            <div className="dmy">
+              <div className="current-time">
+                <Clock format="HH:mm:ss" interval={1000} ticking={true} />
               </div>
-              <div className="temperature">
-                <p>
-                  {this.state.temperatureC}°<span>C</span>
-                </p>
-                {/* <span className="slash">/</span>
-                {this.state.temperatureF} &deg;F */}
-              </div>
+              <div className="current-date">{dateBuilder(new Date())}</div>
+            </div>
+            <div className="temperature">
+              <p>
+                {weather.temperatureC}°<span>C</span>
+              </p>
             </div>
           </div>
-          <Forcast 
-            icon={this.state.icon} 
-            weather={this.state.main} 
-            onCitySelect={this.searchByCity}
-            currentWeather={{
-              humidity: this.state.humidity,
-              visibility: this.state.visibility,
-              windSpeed: this.state.windSpeed
-            }}
-          />
-        </React.Fragment>
-      );
-    } else {
-      return (
-        <React.Fragment>
-          <img src={loader} style={{ width: "50%", WebkitUserDrag: "none" }} />
-          <h3 style={{ color: "white", fontSize: "22px", fontWeight: "600" }}>
-            Detecting your location
-          </h3>
-          <h3 style={{ color: "white", marginTop: "10px" }}>
-            Your current location wil be displayed on the App <br></br> & used
-            for calculating Real time weather.
-          </h3>
-        </React.Fragment>
-      );
-    }
+        </div>
+        <Forecast 
+          icon={weather.icon} 
+          weather={weather.main} 
+          city={weather.city}
+          onCitySelect={searchByCity}
+          currentWeather={{
+            humidity: weather.humidity,
+            windSpeed: weather.windSpeed
+          }}
+        />
+      </>
+    );
   }
+
+  return (
+    <>
+      <img src={loader} style={{ width: "50%", WebkitUserDrag: "none" }} alt="Loading" />
+      <h3 style={{ color: "white", fontSize: "22px", fontWeight: "600" }}>
+        Detecting your location
+      </h3>
+      <h3 style={{ color: "white", marginTop: "10px" }}>
+        Your current location will be displayed on the App <br /> & used
+        for calculating Real time weather.
+      </h3>
+    </>
+  );
 }
 
-export default Weather;
+export default CurrentLocation;
